Other users profile

+++ back/src/user/profile/photos/photo.service.ts
@@ -69,7 +69,7 @@ class PhotoService {
       .del()
       .returning('*');
     console.log('toremove: ', toRemove);
-    if (!toRemove.path.startsWith('src/')) {
+    if (toRemove && !toRemove.path.startsWith('src/')) {
       await fs.unlink(toRemove.path);
     }
   }
